Summary
Hey, I'm back.

It turns out we have a problem when the current user tries to delete any of their own recipes. When he tries to do this thing, all saved recipes up to Mometna are deleted from the database, as well as his profile and data from the database, too.

What I do is delete everything related to this process so far and rebuild it.